fix some casting warning

+++ Figures/CLine.cpp
@@ -17,7 +17,7 @@ CLine::CLine(Point p1, Point p2, GfxInfo FigureGfxInfo)
 	}
     
 	ID = 2200 + newID++;
-	length = sqrt(float((start.x-finish.x)*(start.x - finish.x))-float((start.y - finish.y)*(start.y - finish.y)));
+	length = int( sqrt(float( (start.x-finish.x)*(start.x - finish.x) )-float( (start.y - finish.y)*(start.y - finish.y) ) ) );
 }
 
 void CLine::Draw(Output* pOut) const
